remove duplicate wake-compute loop
